Cache private search provider's template url data
fix https://github.com/brave/brave-browser/issues/28360

Cached default private search provider is added to list
when it's not included in default provider list.

So far Brave is set when previous default one is not included in
the default list. We should set previous one as a default
when new default list is changed and it's not included.