Extend Recover API so that we can update media source for items (#17035)
* Extend Recover API so that we can update media source for items

The media source url could be easily expired. So we might have got to
update media source url before we try recovering the media file. In case
we can notice that the url would be expired from url, we can update media
source first. But mostly we can't infer if its expired or not. So when
we fail to recover media file, we try updating media source by default.

In order to implement this new behavior, also refactored
PlaylistMediaFileDownloadManager. The delegate pattern is extracted out
to callbacks so that we can remember |update_media_src_and_retry_on_failure|
param,