Bump two yanked rust dependencies.
The versions of `cbor4ii` and `ed25519` have been yanked by their
maintainers, indicating they should not be used for future work.

Update to the latest compatible releases to address the audit
warnings.

Result of running `cargo update -p cbor4ii ed25519`.

The `ed25519` 1.5.3 release should be api compatible with our use
and guards against building with the `signature >= 2.0` crate
which it is not compatible with.

The `cbor4ii` 0.2.14 release fixes an issue with `skip_exact`
not returning EOF when passed an empty buffer. It looks like
this isn't called by any current code.