Removing bug-prone base::WriteFile calls
This PR removes the use of the overload in upstream chromium for
base::WriteFile, which happens to be bug prone, due to the way the
return logic has to be checked.

For instance, the check for base::WriteFile in the installer code was
wrong, as the failure sentinel for the bug-prone version is -1.