[brave-news] Feed item parsing clean up
This PR introduces an IDL file to handle `base::Value` checks for feed
parsing. The current code has a number of dereferences in the code that
pose a significant risk, and to guard against them will involve even a
higher number of awkward control flow, which can be quite easy to get
wrong.

Another PR will follow this one converting other parsts of the feed to
also use a type generated by the json-schema-compiler for
parsing/validation.