Fix brave/brave-ios#6294: Bookmark import not working (brave/brave-ios#7002)
While accessing document URL from UIDocumentPickerViewController to access the file startAccessingSecurityScopedResource should be called for that URL.

Given an NSURL created by resolving a bookmark data created with security scope, make the resource referenced by the url accessible to the process. When access to this resource is no longer needed the client must call stopAccessingSecurityScopedResource. Each call to startAccessingSecurityScopedResource must be balanced with a call to stopAccessingSecurityScopedResource (Note: this is not reference counted).

The difference from brave/brave-ios#6781 is state changes causes stopAccessingSecurityScopedResource is being called early so this PR is a temporary fix until brave-core PR will move start and stop secure scope in utility methods in BookmarkExportImport API