Satisfying non-virtual destructor warnings
Brave has overrides for TranslateUiDelegate, which tag certain functions
as virtual, while at the same renaming, and inheriting from the original
class. This means a class with virtuals and yet no virtual destructor,
for which clang has warnings for.

This change marks our override class as final, as the override class has
taken up the name of the original one, and there's no chance anyone can
refer to the original directly, outside the overriding files.