Removing dependency to //components/browser_watcher
The code being removed in the PR was already not in use from its
introduction, as the `ExitCodeWatcher` would never be initialised in the
first place.

Additionally, `//components/browser_watcher` is getting deleted upstream
in `cr113`, while `ExitCodeWatcher` code is getting moved into
`//chrome`, which we can't easily depend on from components.