[json-schema-compiler] Support none as optional
The introduction of the schema compiler has broken brave news, as our
server-side services in some cases provide absent fields with the `null`
value in json, rather than just removing the key entirely, as it is the
case with WebIDL, etc.

This PR adds support to `null` values as a form of absent field marker
as well. This is is being guarded behind a namespace attribute, namely,
[none_as_absent_optional], to make sure chromium codegen is not
affected.