Providing version info as constexpr
This change updates `//brave/components/version_info` to produce
versioning strings through a generated header. This simplifies the
inclusion of the files all around, and remove risks related to the use
of compile time defines, and gn configs carrying those over.

This is being done to make it ease off the changes necessary to give
`GetBraveVersionNumberForDisplay()` a `constexpr` declaration, which will
necessary at some point, due to upstream changes in `cr113`.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/f630755c514e612256c787f5a73a14775c9bccd7

commit f630755c514e612256c787f5a73a14775c9bccd7
Author: Will Harris <wfh@chromium.org>
Date:   Tue Mar 14 00:00:19 2023 +0000

    Make several version_info functions constexpr

    This CL makes most functions in version_info return constexpr
    values.

    This allows them to be assigned to constexpr values elsewhere
    in the code.