Asset ratio parsing with idl (#17558)
* [CodeHealth] Parse asset ratio TokenInfo in WebIDL

This change introduces the use of the json schema compiler to provide
types for the parsing of TokenInfo, in order to simplify the code around
it, making it more robust, and also avoiding refactor changes due to
disrupting `base::Value` refactoring.

* [CodeHealth] Parse asset ratio CoinMarkets in WebIDL

This change replaces the parsing of the json value produced for
CoinMarket from handcrafted base::Value uses, to codegen produced by the
json-schema-compiler. This will also remove the deprecated uses of
base::Value in the code involved.