Remove exception control from challange bypass
The code in challange bypass was attempting to emulate exceptions by
having an error stored in a global local thread state. This is
undesirable, and we should prefer to have failure signaled through
constructs like `base::expected`.

This commit adjusts the return types of all functions that are expected
to fail, and also deletes all exception related code.