Support selected network per origin
We use different pref(kBraveWalletSelectedNetworksPerOrigin) to store that info,
the original kBraveWalletSelectedNetworks will be a fallback pref when
origin is not available when setting/getting selected networks.
Previously cached `chain_ids_` and `network_urls_` are
removed because `JsonPrefStore` already has a layer of memory cache so we won't
directly read from disk.