Fix: Subsequent delta updates on Windows failed
The reason for the bug was that GetPrevVersion in create_installer_archive.py
extracts the previous version's chrome.dll file *into the staging directory*.
This has the side effect of placing the .dll into chrome.7z, which serves as the
target during delta update generation. When the (first) delta then gets applied,
the old .dll gets placed in chrome.7z in Brave's installation directory. When
the second delta then runs, it sees an unexpected hash of chrome.7z and fails.

This commit replaces the implementation of GetPrevVersion with one that does not
have the unwanted side effect.