Pass anti-targeting data by ref
We are changing the way how subvision target data, and anti targeting
resources are being passed around, to have them passed by reference. The
use of pointer allows for nullability, which is not really something
that is supported.

Additionally, `AntiTargetingExclusionRule` is being changed to keep the
reference to the anti target resources, rather than doing a copy of it
during construction.