Replace ParsingResult with base::expected
This change replaces the use of ParsingResult with base::expected. This
makes the signalling of failures easier to read, and avoids in-between
states where you may have no error, and no result.

The types returned by result are now being passed in stack, rather than
as heap objects. This is being done to avoid unnecessary allocations,
and also to reduce the complexity when reading the code.