Show tab search bubble around tab search button (#17973)
On Windows, the button could be on the frame, not on
the vertical tab strip. Currently, when a shortcut for
the search bubble is pressed, the bubble shows up
next to the vertical tab strip even though the button
isn't there. Thus, add a condition to check the button
is on the frame.