Remove `ParsingResult` in Ads (#17905)
* Pass anti-targeting data by ref

We are changing the way how subvision target data, and anti targeting
resources are being passed around, to have them passed by reference. The
use of pointer allows for nullability, which is not really something
that is supported.

Additionally, `AntiTargetingExclusionRule` is being changed to keep the
reference to the anti target resources, rather than doing a copy of it
during construction.

* Replace ParsingResult with base::expected

This change replaces the use of ParsingResult with base::expected. This
makes the signalling of failures easier to read, and avoids in-between
states where you may have no error, and no result.

The types returned by result are now being passed in stack, rather than
as heap objects. This is being done to avoid unnecessary allocations,
and also to reduce the complexity when reading the code.