[CodeHealth] Ledger reference fields as `raw_ref<>`
This change replaces the use of naked pointers for ledger as a class
field a `raw_ref<>`, in line with hardening guidelines from upstream.
Additionally, this change also corrects the passing around of the
reference to be passed by ref, rather than as a pointer, as this makes
the contract clearer for the types holding the reference.

This change is being is required to build brave against M114, as
hardening checks with the clang plugin are being enabled for the Linux
build.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/91cd0d4cb8b3e5ed2ef94bb75fb36d652c478e4a

commit 91cd0d4cb8b3e5ed2ef94bb75fb36d652c478e4a
Author: Keishi Hattori <keishi@chromium.org>
Date:   Thu Apr 6 05:36:18 2023 +0000

    Enable check_raw_ptr_fields for Linux

    Also add enable_check_raw_ptr_fields gn arg so that the check can be disabled by users and for third party libraries.

    This is the third attempt to enable check_raw_ptr_fields for Linux after  r1125241.

    Bug: 1430536, 1273182