Handle kSystemProfileSelectionDefaultNone feature being enabled
This feature is now enabled by default as of cr114. When enabled, the default
profile selection for the system profile is ProfileSelection::kNone to ensure
that no unnecessary services are initialized.

The primary fallout for us involved several calls that were being made to
HostContentSettingsMapFactory::GetForProfile when the system profile was
active (via the ProfilePicker dialog). Call sites assumed that the map was
non-null, which is no longer a valid assumption for the system profile.

Fix follows the approach used in
https://chromium-review.googlesource.com/c/chromium/src/+/3841761

Chromium change:

https://source.chromium.org/chromium/chromium/src/+/271a0a801eb75c53711db785a7979baa88bde63c

commit 271a0a801eb75c53711db785a7979baa88bde63c
Author: Ryan Sultanem <rsult@google.com>
Date:   Tue Mar 28 08:30:06 2023 +0000

    [SystemProfileSelectionDefaultNone] Enable feature by default

    Feature has been at the 1% stable for more than two weeks with a stable
    status.
    UMA stability is good https://uma.googleplex.com/p/chrome/variations?sid=c64b877df26cf72b69e6706f13a49fd5.

    No bugs or crashes were reported since the DEV stage of the feature
    rollout.

    Bug: 1284664