Add a regression test for the freetype crash
The Brave 1.50.121 release crashed loading certain pages on
Linux. Add a regression test based on an isolated reproduction.
The issue seemed to be using the Google Noto Sans VF shipped
as a system font by Fedora, Arch, and a few other distributions
inside a `<select>` element.

This adds a test page, and loads it as part of a browser test.
The test page is minified from https://brave.com/ which also
showed the issue. The font in question was reduced in fontforge
to just lowercase latin glyphs, with all additional tables
removed so save space. The result was included directly in
the test page using a `data:` url.

Resolves https://github.com/brave/brave-browser/issues/29957