Introduce new logic to obtain keyring Ids
The main issue is that so far the keyring Ids were the same for mainnet and testnet for a given coin type. With the introduction of Filecoin support the keyring Id changes depending on the Filecoin network (Failcoin Mainnet and Filecoin Testnet). The solution is to rely on the chain Id and not on the coin type. The method `getKeyringForCoinType` has been renamed to `getKeyringForEthOrSolOnly` as it supports only Ethereum and Solana.
Dapps are only for Ethereum and Solana, so the method can be safely used in the `DappsModel` class. In case of wrong usage it will log and error message, but it won't crash (safer approach to avoid bad UX),