Use the full Noto Sans VF font
Fix css errors which were preventing the page from using the
minimized font from the data: url. Unfortunately this revealed
that the minimized version doesn't actually reproduce the problem;
it was still falling back to the system font.

The should be a stand-alone reproduction, at the expense of an 844 kB
test file.