Fix brave/brave-ios#7403: Fix ServerTrust notification to update the URL bar (brave/brave-ios#7420)
Server Trust does not post a notification if the trust itself hasn't changed. Problem is that the trust on NTP is the same as on an HTTP site. So visiting an HTTP site takes the same trust as on NTP, which is wrong. To fix this, we post a notification ourselves to let us know that the URL changed but the trust stayed the same.
This in turn causes the revalidation to happen.