Remove "brave_ai_chat_endpoint" build argument on Android
Fixes brave/brave-browser#30610

This commit fixes warning

```
The variable "brave_ai_chat_endpoint" was set as a build argument
but never appeared in a declare_args() block in any buildfile.

To view all possible args, run "gn args --list <out_dir>"

The build continued as if that argument was unspecified.
```