[Android] Changes for CachedFlag IsEnabled method
Chromium change:
https://github.com/chromium/chromium/commit/1ce30b46f0d225c5063d0db14d423df8d8ea230e

[Android] Move lists of CachedFlags from ChromeCachedFlags
ChromeFeatureList now contains the lists of CachedFlags (and the
CachedFlags themselves).

It is necessary to have a list of all CachedFlags in
ChromeFeatureList, since we need to map the feature name to the
legacy shared preference key in the CachedFlag when:
1. Setting up test values via @EnableFeatures/@DisableFeatures
    * Currently done through CachedFeatureFlags#setForTesting, but
      requires CachedFeatureFlags to know legacy SharedPrefs key
      overrides, which blocks its migration to //base.
2. Called via Java_CachedFeatureFlags_IsEnabled()
    * Currently uses the default from native passed as a param through
      JNI.

This also adds a test to ensure flags aren't declared and not cached,
which is a common pitfall.

Bug: 1442347