Use EvalJS without `EXECUTE_SCRIPT_USE_MANUAL_REPLY` (#18754)
This PR changes the use of EvalJs in a similar fashion done upstream,
removing on the uses of `content::EXECUTE_SCRIPT_USE_MANUAL_REPLY`, and
modifying the scripts used in tests to return values directly, rather
than relying on `domAutomationController` to transmit back results,
which can be finicky.

With the new approach, values are returned through promises, or directly
as a return value.