Changes ChromeSwitchPreference to ChromeBaseCheckBoxPreference class for sync data types screen on Android.
Related Chromium changes:
https://source.chromium.org/chromium/chromium/src/+/29ac6dc2ea650d1768a9b222e6c49accf8100f19
[Signin][Android] Use a map for mSyncTypePreferences.

This is just a refactoring CL and will not change any logic.

Fixed: 1407184
Change-Id: I92a04692af00aefc6654ab9b681c9efae914ab10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4422769

and

https://source.chromium.org/chromium/chromium/src/+/b3d30106aa25f614cfb46524e40972f9130573d5
[ManagedSyncSettings] Sync Settings on Android Reflect Managed State

When sync is enabled and SyncTypesDisabledList policy is set with some
data types, the checkboxes are not disabled. However, checking the
boxes did not have any downstream effect.

This CL sets the managed preferences when sync types are disabled by
the SyncTypesListDisabled policy.

Bug: b/277005858
Change-Id: Ibdb031885918e475972b94f488ac1e8f317372d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4507582

Fixes brave/brave-browser#30964