P3A: Accept multihash nitro enclave user_data keys
Version 1.3.1 and earlier of the nitriding-daemon report the TLS
certificate fingerprint as part of the AWS-signed attestation
document in a form that's awkward to part. Accept a multihash in
the `user_data` field as well as the older `sha256:<binary digest>`
format so we can transition to it on the server side.