[Android] New signature for SigninManager.create
Chromium change:
https://github.com/chromium/chromium/commit/2e8401354bf3451872450caac741bc87fe84d385

[android] Remove some calls to deprecated SyncServiceFactory.get()
Hopefully no behavior changes. These calls were underlying calls to
getLastUsedRegularProfile(). Instead, call
SyncServiceFactory.getForProfile() and pass an explicit profile
reference. The references are chosen from one of the 4 options below.
The focus is not to accidentally replace the regular profile with the
incognito one.

1. A nearby field/variable which is still getLastUsedRegularProfile()
if you follow the stack trace.
2. Use of TabModelSelectorProfileSupplier + getOriginalProfile().
3. The profile in SigninManagerAndroid, since that service is not
instantiated in incognito.
https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:chrome/browser/signin/signin_manager_android_factory.cc;l=16;drc=b35d0400e78c4216ea3e6941d91373147572953c
4. The profile obtained from the WebContents. This only happens in
AddressEditor.java, where the reference can come from
SaveUpdateAddressProfilePromptViewAndroid. In this particular case there's no need for getOriginalProfile(), because the save prompt is
not shown in incognito.
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/android/autofill/save_update_address_profile_prompt_view_android.cc;l=57;drc=74077ae07468070edcaac721aef246dd5fcf8092

This CL introduces new SyncService fields/arguments in some layers.
We add @nullable according to whether the existing code handles a
null object, not whether the object can really be null.
Fixing null support is off the scope of this CL. The null SyncService
story is also not ideal IMO (rant, https://crbug.com/1434356#c8)

Disable-Rts: True
Bug: 1380925