Fix brave/brave-ios#7663: Dismiss popup in playlist onboarding when tapping Maybe Later (brave/brave-ios#7664)
This moves the dismissal environment call to the parent View. This is likely a side-effect of path taken to render the same buttons in one view in another like so:

```
DetailsStepView(advance: {
  model.step = .tryItOut
})
.expandedViewButtons
```

Since the View itself (`DetailsStepView`) is not used, and thus has no `body`, it can't access the environment.