Refactor on ramp list parsing to new format.
Instead of separate files for on ramp and off ramp tokens, each with a
separate list of supported tokens for each provider, combine all into a
single list of tokens, with new fields "on_ramp_providers" and
"off_ramp_providers" which list which on ramp and off ramp providers
support that token.

This commit updates how the lists are parsed, but the internal
representation remains the same.