Add all Leo skia colors to light and dark color mixers
This makes referring to the colors from UI classes easier by not having to redefine a new variable with a leo dark or light equivalent. It takes advantage of the fact that all Leo colors already have light and dark equivalents. UI that wishes to have different Leo colors for light and dark can and will still need to specify that explicitly.