Keep HideBottomViewOnScrollBehavior (#19764)
Keep `HideBottomViewOnScrollBehavior` class that was removed by Proguard optimizations. 
This class is referenced in an XML layout under the attribute `app:layout_behavior`, in particular in the main Brave Wallet layout (`activity_brave_wallet.xml`). 
Proguard is not able to determine whether this class is used and was removing it in release builds only.