Avoid dangling pointers in `BravePrefProviderTest` (#19799)
When using `ScopedDictPrefUpdate` to create dictionaries, a raw ptr is
used to keep track of it. However, these prefs end up getting destroyed
when these tests eventually call
`MigrateShieldsSettingsFromResourceIds`.

This change makes sure that the `ScopedDictPrefUpdate` instance is
dropped as soon as it is not needed anymore, to avoid the internal
pointer in that instance to dangle, as the pref it is referring to gets
destroyed.