Fix brave/brave-ios#7909: Handle JitsiMeet delegate methods to close unjoined calls. (brave/brave-ios#7910)
The `ready(toClose:)` delegate method is called by the JitsiMeet SDK when you try to leave a call that hasn't officially "joined" yet. It is also called in other situations though like terminating an active call or when the connection is halted and needs to reconnect the call so we will only dismiss the `JitsiMeetView` when the user hasn't joined a call yet.