Avoid dangling pointer in `speedreader` (#19827)
This is a fix for a few cases where `raw_ptr` instances were left
dangling during unit tests.

The reader FFI pointer passed between rust and cxx has a quirky
lifetime, and had to be explicit released for destruction on the rust
side. Additionally, this change corrects a dangling profile pointer in
one of the tests.