Split EthereumProviderImpl into EthereumProviderHost and EthereumProviderService.
Each renderer has one EthereumProviderHost which talks to keyed service
EthereumProviderService. When binding remote in EthereumProviderHost to
EthereumProviderService, BraveWalletProviderDelegate will be provided
which will be stored in EthereumProviderService as a map keyed by
ReceiverId.

This is to share EthereumProviderService for both requests initiated
from renderer and native UI which can be used for future Wallet Connect
feature.