Update cargo-audit to v0.18.1
Work around a compatibility issue between our previous version and
the chromium 117 rust toolchain, probably from incompatible nightly
versions, but updating our vendored source to the latest release.

This is the contents of the published package downloaded and unpacked
from https://crates.io/api/v1/crates/cargo-audit/0.18.1/download
with the README.chromium updated manually.

We build this in CI with cargo, so we updating other vendored deps
is not necessary.