[Android] Changes for omnibox suggestions
Chromium change:
https://github.com/brave/chromium/commit/ba278ecef32b522a509166fa04052503d3053b0e

Dropdown Builder: Introduce support for horizontal render groups.
This change enables DropdownItemViewInfoListBuilder to create
Horizontal Render groups.

Horizontal render groups are constructed with the help of a single
PropertyModel, that is internally turned into a desired view hierarchy
by the appropriate component.

A single PropertyModel is created for the entire group, and amended
with every next suggestion that belongs to the same group.

Bug: 1474087