speedreader: port to kuchikiki 0.8.2
We renamed our fork of the kuchiki tree-handling crate since
upstream is unmaintained. It's easier to build against released
crates now that we're vendoring, so move to the new publication
crate name. This also addresses a `cargo audit` warning about
the unmaintained dependency.

Note however that we're still using a brave-specific fork from
git here, so the crate itself is still installed from DEPS.

This adds a new transitive dependency on `indexmap`, which is used
to roundtrip html attributes with stable ordering so unit tests
don't need trivial updates whenever we touch something.
This dependency was for a while available in upstream chromium,
but has since been dropped, so we're using our own copy under
brave/third_party/rust.

kuchikiki 0.8.2 specifies newer versions of its other dependencies
but seems to still work when built against the old releases, so
those are left to another commit.

This replaces the DEPS entry with a copy of the source from
https://github.com/brave/kuchikiki for simpler updating.
Note that this is still a patched version, different from
the 0.8.2 package published on crates.io.