Show vertical tab strip on mouse hover in Browser fullscreen (#20230)
* Show vertical tab strip on mouse hover in Browser fullscreen

We were showing vertical tabs on browser fullscreen. But we've got
mixed response about the visibility. Some want it to be visible on the
browser fullscreen mode, while others want the opposite.

So we're going to hide vertical tabs by default but support revealing it
by mouse hovering. And especially on Mac, there's a pref named
"Always show toolbar in fullscreen". Vertical tab strip will be applied
that pref too so that users can decide the default visibility.