Prefe constexpr over static data
This change removes the use of static data in certain cases where it is
not required, in order to avoid `exit-time-destructors` warnings. These
warings are emitted to indicate that the destruction order of a static
object is indeterminate, and therefore unreliable to use during
termination.

With this change, these sites are being converted to simple constexpr
data.