[Android] Temporarily disable java lint checks
There is a known issue of mismatching the newest depot_tools and older Chromium versions (older than C119) https://bugs.chromium.org/p/chromium/issues/detail?id=1462204#c15
The problem is that requirement for google-java-format is already in depot_tools, but DEPS for it added only in C119.
So we need temporarily disable it to unblock PRs on older Chromium versions.