Remove `SidePanelCoordinator` `final` keyword substitution
`SidePanelCoordinator` is tagged upstream as `final`, and as a
workaround to be able to derive from it, an override had been introduced
at some point to replace the keyword `final` with naught. This type of
redefinition is unreliable, as it is UB, and it also becomes highly
viral.

This change replaces the substitution with a patch. The substitution was
removing the `final` keyword from the signature of other functions in
different headers, and leaving it with no token to indicate the function
is an override, causing a build failure.