Make `AdType` a class enum type
AdType is a class with a single enum scoped into it, which it makes it
functionally a class enum. This change converts the class to an enum
class, to make of what is offered by the language out-of-the-box.

Additional conversion functions have been provided, in a similar fashion
how upstream provides such functions for conversion to/from char, as
well as with a mojom paired enum.