Declare sorted fixed `flat_map`/`flat_set`
Sorting in these containers on a constexpr context is an expensive
operation for the compiler, which increases builds time, but the cost
can become so high to the point of the compiler giving up on a
`constexpr` evaluation for the declaration, which would cause a build
failure.

Errors of this nature have already happened on `canary` possibly due to
recent upstream changes using concepts, which might be more onerous to
evaluate.

This change updates most of the use of `MakeFixedFlatSet`, and
`MakeFixedFlatMap`, to `MakeFixedFlatSetSorted`, and
`MakeFixedFlatMapSorted`, to avoid the expensive compile time sorting
computations with the previous counterparts.