[ui] Add metadata to view classes
Chromium has added a validation to `View::AddChildView` and its
variants, constraining its uses to objects that have this matadata
included.

This change adds metadata to all view class in the codebase that are
required to satisfy this requirement.

An upstream patch has been submitted as well, adding metadata to
`ReadLaterSidePanelWebView`, to permit the dervied types in Brave to
also declare metadata. https://crrev.com/c/5040953.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/080f886cad2cdd0a0db2e1534e1a8ed6a9fcc923

commit 080f886cad2cdd0a0db2e1534e1a8ed6a9fcc923
Author: Allen Bauer <kylixrd@chromium.org>
Date:   Wed Nov 15 17:34:29 2023 +0000

    Guard View::AddChildView (and variants) to ensure view class has metadata.

    Bug: 1485768