[Android] Fix for unregistered kEnableSnippetsByDse preference
kEnableSnippetsByDse preference is only registered when Feed is enabled, however in Chromium they try to set it even when Feed is disabled.
Since we don't use Feed in Brave we need to disable this code.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/1f4c1c5e28a37e3104c015eb30593d92cb3c516e

[DSE] Identify EEA countries

In this CL, we expose an API to identify EEA country to Java in
TemplateUrlService. Add a parameter and update chrome://flags.

Bug: 1475011