[canary] Initial basic tooling for daily canary
This is a simple script, only a preliminary idea, experimenting on ways
to automate nearly all tasks relating to daily canary. The current
script should serve as a simple prototype to explore ways we could
interact/change with the process of rebasing chromium in general.

This initial version only supports updating a version, and going through
the three steps of 1) updating `package.json`, 2) updating patches, and
3) updating the resource files.

Further work has to be done how to better interact with the who
initialisation process, and how to handle the "conflict-resolved" step.

To run the script, just use something like:

```
script/version_up.py --previous=origin/cr120 --to=121.0.6138.1
```

At the moment this has to be done from the root path for brave-core, but
this will be eventually corrected.