[rewards] `ContributionPublisherPairListCallback` with `base::Bind`
This change updates `ContributionPublisherPairListCallback` to use
`base::OnceCallback`, as `std::function` is banned, and doesn't provide
the same strict requirement for captures that is demanded by
`base::Bind`.

This change is part of a batch of changes to allow us to use clang-tidy
instrumentation across the codebase, during PR submission.