[rewards] `GetTransactionReportCallback` with `base::Bind`
This change removes the use of `std::functon`, and `std::bind` for
aroud `GetTransactionReportCallback`.

These functions are banned in chromium [1][2]. For the most part,
`base::OnceCallback` has more strict validation around captures, which
vanilla lambdas, and `std::bind` lack.

This change will also allow us to leverage the dangling pointer
machinery to identify issues, as even `base::Unretained(this)` uses
dangling pointer instrumentation, which is an extra protection when
compared to a naked lambda capture of `this`. Additionally, due to the
nature of how base::OnceCallback has to be converted to an rvalue
reference to be run, this allows us to use static analysis to detect
bugs involving double calls of certain events.

[1] https://chromium.googlesource.com/chromium/src/+/main/styleguide/c++/c++-features.md#std_bind-banned
[2] https://chromium.googlesource.com/chromium/src/+/main/styleguide/c++/c++-features.md#std_function-banned