[Android] Replace TabCountProvider with a supplier
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/5f906a6d1539a1902eabd949c79822a1751c1505

Replace TabCountProvider with a supplier

TabCountProvider can be replaced by directly observing the
TabModelSelector's getCurrentModelTabCountSupplier().

Where the TabModelSelector is already propagated we don't need a
separate set method for the new observable supplier so this reduces the
method count by quite a bit.

In places without a TabModelSelector the ObservableSupplier<Integer> can
be set in place of the old TabCountProvider

Bug: 1499464