[Android] TabImpl turned to package private
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/77c918240217600b4592de8746308ec05f634b22

Android: Turn TabImpl to package private

TabImpl is an implementation detail of Tab interface, and therefore
the access from other packages needs to be discouraged. This CL
makes it package-private.

A few methods of TabImpl are still required by external test classes.
They are accessed via TabTestUtils. Moved the remaining methods to
the util class to remove all the direct usage.

 - initialize
 - setWebContentsState
 - setLastNavigationCommittedTimestampMillis

Bug: 995903