Leo text extraction via screenshot text extraction
If text_recongition is available (Windows and macOS) then it will
use the RenderWidgetHostView's surface to capture a screenshot of
the web_contents for certain pages that are canvas based such as
docs.google.com.

It currently also uses this method for screenshots of Twitter for now.
Although it would be possible to remove Twitter in the future when we
support better extraction methods.

The feature is currently limited to only capturing the visible area to
the user.