[Android] Fix for GetLocalPrepopulatedEngines changes
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/0cd7f0c1767b6d5d15a0e99adcb39b15ab76c412

🧇 Ensure that GetPrepopulatedTemplateURLData never gets null prefs

A previous change started making it required to pass prefs when calling
this method, but Android is the last platform where the
SearchEngineChoice feature that requires it is not enabled. In
preparation of rolling out changes related to this feature on Android
too, we close this outstanding issue.

Noting that:
1. We still have many tests passing null prefs, so this
exception is allowed for now, until we fix this in follow-ups, as
part of https://crbug.com/1499181
2. We are assuming that the service should be available when methods
are called from Java, and currently return fallback values if it's
not the case. We will move the service availability check to Java in
another follow-up (see b/318339172).

Bug: 1499181
Fixed: b:303632061