Enable Widevine on all Linux archs (#21040)
On Arm64 Linux, Widevine is not available via component updates. But
distributions such as Raspberry Pi OS ship it. So users have ways of
obtaining the library. This commit makes the browser attempt to load
Widevine on non-x64 Linux architectures (and thus on Arm64). Users can
then install it manually to see DRM-protected content.