[Android] Site settings fixes
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/93a3620ca2272ce0352112ebe0f80e6ccbb4b73f

Add a ContentSettingsType enum defined in mojo

The goal of this change is to migrate the source of truth for the
ContentSettingsType from a C++ header into mojo.  Generated C++ from the
mojo definition is exactly the same as the enum in
content_settings_types.h, and moving the source of truth to mojo lets us
generate a TypeScript equivalent enum more easily.

This change also removes the complexity of type conversion with
`EnumTraits<>`.  Avoiding use of `EnumTraits` is recommended in the
style guide.
https://chromium.googlesource.com/chromium/src/+/HEAD/docs/security/mojo.md#using-mojo-enums-directly-when-possible

Generated Java enums are slightly different, unfortunately requiring a
small change to the type name in every usage.

Bug: b:317035343