Explicitly enable std on the indexmap crate
The `indexmap` crate tries to detect with a build script whether
the `std` library crate is available. This fails on our android
integration builds:

    ACTION //brave/third_party/rust/indexmap/v1:indexmap_lib_v1_build_script_output(//build/toolchain/android:android_clang_x86)
    error[E0463]: can't find crate for `std`
      |
      = note: the `i686-linux-android` target may not be installed

However, the check can be overridden by explicitly requesting the
feature, and here we do so in our build. This is the default in the
newer v2 release of the crate, so this change is likely to remain
correct going forward.