Fixes settings content page loading error.
This removes "Performance" item from the advanced items on the page. We
moved Performance from its own page (and path) to living under the
System settings. Because there's no Route.path for this item the page
throws an exception. The item doesn't make much sense on the content
page since we consider it more of a systemwide setting (with the ability
to add individual sites to always keep active, but that's hardly a
reason to show it on the content page).