Remove dangling folder ID from playlist order preference (#21696)
Previously, a data pruning routine was implemented to handle new data schemas.
However, this routine was flawed and left data in the order preference,
leading to crashes. To resolve this, migration code has been added.

As this feature is nearing release on Android and Desktop platforms,
data pruning will no longer be necessary. Therefore, the pruning code
has been removed.