[Android] Remove getLastUsedRegularProfile from UpdateMenuItemHelper
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/26ca76247ab0d8c08cf65894f1686dabe8dd6e08

Remove getLastUsedRegularProfile from UpdateMenuItemHelper.

This converts the singleton instance to a Profile-keyed
class.

This also renames shouldShowUpdateBadge to canShowUpdateBadge
as this was confusing during reading. When reading, should
(to me) implied that the item was to be displayed if true,
but it is only shown based on other criteria like being out
of date. Thus, "can" more clearly read to me that the update
badge is allowed to be shown if other conditions are met.

Bug: 1410601