speedreader: specify the correct html5ever version
kuchikiki v0.8.2-speedreader declared a dependency on
html5ever 0.26, but we actually build against 0.25.1.
Change the declaration here to match what we have
vendored, which works fine in practice.

Bump to v0.8.3-speedreader which declares a dependency
compatible with what we're building against to simpilify
later updates based on the Cargo.toml information.

Resolves https://github.com/brave/brave-browser/issues/35708