Set `UserActivation` before requesting fullscreen
On some sites, such as Vimeo, Rumble and etc, fullscreen button on PIP window
was not working. This was because Fullscreen is allowed only for certain cases.

In fullscreen button on PIP window case,
we can consider this as user activation. So sets `UserActivation` before requesting
fullscreen mode so that we can proceed to fullscreen mode.

https://source.chromium.org/chromium/chromium/src/+/d6b28888aa75d2d8579808102a50a8aceb9afeaf:third_party/blink/renderer/core/fullscreen/fullscreen.cc;l=351