`View::Layout` updated to be called through `View`
Changes in upstream have deprecated calling `Layout()` directly, in
order to have passing through upstream. There are now guidelines of how
to use `Layout()` [1], but the overrall suggestion is to avoid using it
in many cases.

This change corrects all our declaration to use the `PassKey` provided,
replacing direct calls in certain sites with the new deprecated one.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/1e3a1e996ad101e147345cc7e98edc90f99720cb

commit 1e3a1e996ad101e147345cc7e98edc90f99720cb
Author: Peter Kasting <pkasting@chromium.org>
Date:   Thu Feb 1 18:10:41 2024 +0000

    Change View::Layout() to take a PassKey.

    This prevents direct access from outside View itself.

    Must be landed as a single large CL, or LayoutSuperclass() will not
    select the correct method to call.

    Bug: 1521108