Fix override failures around explicit `ScriptPromise` subclassing
Upstream has introduced explicit subclasses for `ScriptPromise`. This
has broken our overrides for `BatteryManager`, which are corrected in
this change.

Additionally, failures have occurred around codegen for some binding and
the use of `RelatedApplication` in it, which was being only
forward-declared, and brave's overrides for `ToPageGraphBlinkArg`. This
change includes the header for `RelatedApplication` in our overrides to
avoid such failure.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/be84811ecc96cd6f2d003d54615f4ed1fd87a3d7

commit be84811ecc96cd6f2d003d54615f4ed1fd87a3d7
Author: Nate Chapin <japhet@chromium.org>
Date:   Thu Feb 1 18:59:51 2024 +0000

    Add explicitly typed subclasses of ScriptPromise/ScriptPromiseResolver