Defer media detection on certain sites (#21332)
* Defer media detection on certain sites

On certain sites, such as YouTube, media detection is resource-intensive.
This is due to the necessity of loading the same page within a background
web contents once again, so that we can hide media source API or fake UA
string. Otherwise, Getting downloadable video url is impossible.

Previously, we executed the media detection script each time navigation
occurred, displaying playlist action icon view on location bar when there're
media that can be added. But this had noticeable impact on performance
and memory usage when visiting sites like Youtube.

To mitigate this, we will defer media detection on these sites. Instead,
the playlist action icon will be displayed when the page URL aligns with
known rule sets. Upon clicking the icon, the media detection process will
be initiated.

Note that on sties in which media detection can be done without a background
web contents, media detection will be executed as before, on navigation.