Fix AIChatCredentialManager Tests
They were failing because the SDKs that were being initialized on the
sdk thread and stored as members on the SkusServiceImpl were not also
being destroyed on the sdk thread. I had made some changes to handle
this case on shutdown, however the SkusServiceImpl::Shutdown function
was not being called in the tests, which led to crashes when the tests
completed.

To resolve this, I just changed SkuServiceImpl::GetOrCreateSDK ->
::CreateSDK, and just stopped storing them on the SkuServiceImpl,
avoiding the problem entirely.

Also, add missing *_task_runners to the SkuServiceImpl in
AIChatCredentialManagerUnitTest.