Update ppoprf source to v0.3.1
Align the vendored source code with the v0.3.1 release specified
by `star_constellation`. We were already calling this version
that in `BUILD.gn` but the source code corresponded to the v0.3.0
release instead.

The only source change is to error handling in the `ppoprf::eval`
function which is not used client-side, so there is no behavior
change to the browser. This just removes a discrepancy so future
vendor diffs will be clean.

Addresses an 'patch...was not used' warning when running `cargo`
against third_party/rust/Cargo.toml.