Add missing README.chromium files in third_party/rust
These were generated running `gnrt` on a branch where that
somewhat works, and then manually edited to remove the new
'License File' field. Upstream is using this because their
README annotations is in a separate part of the source tree
from the actual crate source, but our current style for
vendored crates doesn't have that issue.

core-foundation-sys is marked as not security-critical.
This is used to access the time zone database on macOS,
so the input isn't coming from network sources.