`sql::Recovery` deleted upstream
This class has been replaced with the use of
`sql::BuiltInRecovery::RecoverIfPossible`, which is simpler and more
straightforward. Addtionally
`kAttributionStorageUseBuiltInRecoveryIfSupported` was also dropped as a
result of this change, and it has also been deleted from brave. This
feature is part of attribution reporting, which is already off.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/b3be6371d9bde45b2e1d760836134a52ed05155a

commit b3be6371d9bde45b2e1d760836134a52ed05155a
Author: Evan Stade <estade@chromium.org>
Date:   Tue Feb 27 21:11:09 2024 +0000

    SQL: remove legacy recovery code.

    Built-in recovery has been launched to stable for all databases
    that use recovery (in M122).

    -12KB binary size on Android.

    Bug: 40061775