[Android] New method for ThemeColorObserver
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/2c24369515ffb61dfca8e5ae252240e41b5571e4

[SurfacePolish] Fix bug about toolbar flips to white when navigating
from NTP to webpage

When navigating from the NTP to a webpage, the toolbar color initially
changes to the default toolbar color during the loading phase. It will
change to the browser tab's color or stay unchanged after the loading
phase. The loading process occurs directly on the NTP. However, with the
Start Surface Polish enabled, the NTP's toolbar has a different color
than the default toolbar, and this results in a flip of white when
navigating from NTP to webpage, as shown in the
image:
https://drive.google.com/file/d/1ywogmzelW1OQwGdPCvrd0EndFkAbvCoM/
view?usp=sharing&resourcekey=0-XV6pRmumU3L4vy36HRSJ4w.
This CL aims to fix this problem.

When navigating from the NTP to a webpage, the ToolbarPhone#mVisualState
will immediately turn from VisualState.NEW_TAB_NORMAL or
VisualState.NEW_TAB_SEARCH_ENGINE_NO_LOGO to VisualState.NORMAL called
by function ToolbarPhone#onTabOrModelChanged. This is the starting point
of the loading phase.

The ending point of the loading phase is the time we know if we have a
website theme color. The time we start to know whether the website has
its theme color is the time point ToolbarPhone#onPrimaryColorChanged()
gets called. Since this function won't be called when the website does
not have a specific theme color, I added a function
ToolbarPhone#onPrimaryColorGotUpdateNotified() instead. The newly added
function will be called at the same time if
ToolbarPhone#onPrimaryColorChanged() gets called. And the ending point
is marked inside this function.

After this CL, the navigation process will look like
this:
https://drive.google.com/file/d/1FchVCPWj3DnLaMoZ1Z5Ra_7Z6SIavRSV/
view?usp=sharing&resourcekey=0-tT-y6F12OihXzxjeTnRaIw

Bug: 323888159