[Sync] New COLLABORATION_GROUP ModelType
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/6c87be591459281222cb361f166c3dd1d7327cd1

[Sync] Add scaffolding for a new COLLABORATION_GROUP ModelType

COLLABORATION_GROUP is a read-only (for the client) datatype, that will
be used to invalidate cached People Groups data.

For now it is bundled together with SHARED_TAB_GROUP_DATA, e.g. uses
the same feature toggles and controlled by the same UserSelectableType.
This, especially the later, could be changed in the future in case other
features will ever start to rely on COLLABORATION_GROUP.

Bug: 301389859