Fix intermittent crash in video overlay window
Previously, in order to replace upstream's `VideoOverlayWindowViews`
with our `BraveVideoOverlayWindowViews`, we were relying on
`Widget::Close()`. In normal cases, this would delete the widget pointer.
but in this case, widget is not initialized yet and it left the original
widget pointer dangling. So we must not create the original object at all.