Check Playlist media download callback properly
As the DownloadManager's callbacks are async. So when the callback is
invoked, the current item could be already a new one. In this case, we
shouldn't reset internal data.

This is ad-hoc for fixing problems and we need rafctor the code to make
data's lifecycle would be bound to the lifecycle of download item.