Move `Appearance > Content` settings to the top level of settings menu
As The `Content` setting was hard to find. This PR moves the `Content`
settings to the top level of the settings menu. At the same time,
moves `Playlist` setting into the `Content > Playlist` like `Content > Speedreader`.
These two features are related to contents that're displayed, so it makes
sense to group them together.