[iOS] - Leo force receipt refresh - 1.65.x (#22725)
Fix string showing incorrect yearly vs. monthly.
Add potential fix for payments not working.
Add debug local validation to see what the AppStore sees.

Force receipt refresh in the middle of a transaction!
This seems to give us a receipt BEFORE finishTransaction is called so we can validate the receipt on the back-end before marking the transaction complete.